move asset loader files
