+++ wp-content/themes/labor-day/inc/utils/class-asset-loader.php
@@ -6,7 +6,7 @@
  * @since 2.0
  */
 
-namespace ChoctawNation;
+namespace ChoctawNation\Utils;
 
 /** On Construction, enqueues the assets
  *
